requirements.txt
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.29
langgraph==1.1.6
langchain-text-splitters==1.0.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.31
MarkupSafe==3.0.3
mypy_extensions==1.1.0
networkx==3.6.1
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
pydantic==2.13.0
pydantic_core==2.46.0
pypdf==6.1.2
Pygments==2.20.0
pytest==9.0.3
pytokens==0.4.1
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
ruff==0.15.10
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
Werkzeug==3.1.8
xxhash==3.6.0
zstandard==0.25.0
